Move the logging
Do the logging closer to the source, in flatpak-dir.c
instead of in flatpak-transaction.c. This ensures that
all changes end up in the journal, regardless whether
they are done though FlatpakTransaction or not. For
example, the system-helper is using FlatpakDir api
directly, and will now also log to the journal.