adapt to logging changes
Drop the result column, we no longer log failures,
so no need for this field. Deal with the fact
that the installation and remote fields may now
contain the path of local repo. And handle the
system helper logging on behalf of another app,
by looking for and handling OBJECT_UID and
OBJECT_EXE.